Group the relative path tests into one class beside the mock's tests
